% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_study_functions.R
\name{association_study}
\alias{association_study}
\title{Function to perform associations between two parameters
Associations with p values calculated by permutation, using coin package
Note: dataset is expected to be in "long" format for this function.
Note2: missing values are (silently) deleted}
\usage{
association_study(
  subset.parameter = "Granulocytes_Count",
  subset.column = "Cellname",
  dataset = data.long,
  numerical.value = "Concentration",
  comparison.value = "Frailty.index",
  Stratum = c("Age.category", "CMV.status"),
  n.resample = 100,
  ...
)
}
\arguments{
\item{numerical.value}{a numerical value for the comparison}

\item{comparison.value}{the second parameter. If it is a factor, the wilcoxon test will be used. If it is numerical, the spearman_test will be used}

\item{Stratum}{character string. Can be one or several blocking variables.}

\item{...}{}
}
\value{

}
\description{
Function to perform associations between two parameters
Associations with p values calculated by permutation, using coin package
Note: dataset is expected to be in "long" format for this function.
Note2: missing values are (silently) deleted
}
