% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_study_functions.R
\name{association_study_wide}
\alias{association_study_wide}
\title{association study function that can be used with data in wide format:}
\usage{
association_study_wide(
  response.var = "Granulocytes_Count",
  dataset = data.wide,
  explanatory.var = "Frailty.index",
  stratum = NULL,
  n.resample = 100,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
association study function that can be used with data in wide format:
}
