% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{associations_per_ID}
\alias{associations_per_ID}
\title{Associations per individual}
\usage{
associations_per_ID(x, dataset, comparison.var, nresample = 10^4)
}
\arguments{
\item{x}{Subsets in which comparison should be made}

\item{dataset}{a data frame}

\item{comparison.var}{comparison var (for example Age)}

\item{nresample}{times to resample for permutation test (see coin::approximate)}
}
\value{
data frame with results
}
\description{
This function will investigate the association between biomarkers over time
(longitudinal analyses) with ID/individual as blocking variable.
Note: deprecated function. Can be replaced by test_association
}
