% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_selection.R
\name{BH_selection}
\alias{BH_selection}
\title{BH selection}
\usage{
BH_selection(data, p_value = "p.value", FDR_cutoff = 0.15, hist = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{p_value}{character value of column containing the p values}

\item{FDR_cutoff}{numerical value. FDR cutoff}

\item{hist}{logical. Whether or not to create a histogram}
}
\value{
data frame with added columns with FDR and FDR selection
}
\description{
Function to create a selection based on a set False Discovery Rate
I used this function in combination with the association_study function.
}
