% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbbPalette}
\alias{cbbPalette}
\title{Colorblind friendly color palette}
\format{
An object of class \code{character} of length 8.
}
\source{
http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/
}
\usage{
cbbPalette
}
\description{
Palette with black.
}
\keyword{datasets}
