% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_low_group_numbers}
\alias{check_low_group_numbers}
\title{Check low group numbers}
\usage{
check_low_group_numbers(data, stratum = NULL, response.names = NULL)
}
\arguments{
\item{data}{data.frame}

\item{stratum}{any blocking vars. can be empty}

\item{response.names}{any response names/other vars to block by. Can be empty}
}
\value{
no return, unless too low observations, then an error will occur
}
\description{
Check whether number of observations are too low in some groups and return an error if so
}
