% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_observations}
\alias{count_observations}
\title{count observations}
\usage{
count_observations(data, stratum = NULL, response.names = NULL)
}
\arguments{
\item{data}{data.frame}

\item{stratum}{any blocking vars. can be empty}

\item{response.names}{any response names/other vars to block by. Can be empty}
}
\value{
a data frame with number of observations
}
\description{
Count observations per group in an association study.
Output is a data frame, with the strata with smallest number of observations on top
}
\examples{
count_observations(immune_data, stratum = c("Batch", "Sex"))
}
