% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create.matrix}
\alias{create.matrix}
\title{function to create FDR or rho matrix for input in the corrplot function of
the package corrplot}
\usage{
create.matrix(
  data = correlation.results,
  x = "first.var",
  y = "second.var",
  values = "rho"
)
}
\arguments{
\item{values}{}
}
\value{

}
\description{
function to create FDR or rho matrix for input in the corrplot function of
the package corrplot
}
