% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Create corrplot matrix
function to create FDR or rho matrix for input in the corrplot function of
the package corrplot
Expects input from p values/FDR values or rhos, in format as table output from
association studies (see function association_study) (correlation results)}
\usage{
create_matrix(
  data,
  x = "first.var",
  y = "second.var",
  values = "rho",
  markers.to.test
)
}
\arguments{
\item{data}{input data (data frame or matrix)}

\item{x}{first variable}

\item{y}{second variable}

\item{values}{values to use}

\item{markers.to.test}{all biomarkers to test}
}
\value{
matrix for input in corrplot function
}
\description{
Create corrplot matrix
function to create FDR or rho matrix for input in the corrplot function of
the package corrplot
Expects input from p values/FDR values or rhos, in format as table output from
association studies (see function association_study) (correlation results)
}
