% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{geomean_conf}
\alias{geomean_conf}
\title{Geomean + Conf interval
Function to return geomean values with 95\% confidence interval}
\usage{
geomean_conf(x, n.digits = 2, na.rm = FALSE, conf.level = 0.95)
}
\arguments{
\item{x}{numeric value}

\item{n.digits}{digits to preserve in output}

\item{na.rm}{whether or not to remove NA's}

\item{conf.level}{confidence interval, standard 0.95}
}
\value{
string with geomean and confidence interval
}
\description{
Geomean + Conf interval
Function to return geomean values with 95\% confidence interval
}
