% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{geomean_conf_2}
\alias{geomean_conf_2}
\title{Improved version of geomean_conf for latex tables:
Inspired by this post: # https://stackoverflow.com/questions/44325464/how-to-control-knitr-kable-scientific-notation}
\usage{
geomean_conf_2(
  x,
  n.digits = 3,
  na.rm = FALSE,
  conf.level = 0.95,
  scientific.above = 10000
)
}
\arguments{
\item{x}{numeric value}

\item{n.digits}{digits to preserve in output}

\item{na.rm}{whether or not to remove NA's}

\item{conf.level}{confidence interval, standard 0.95}

\item{scientific.above}{threshold above which scientific notation should be used}
}
\value{

}
\description{
Improved version of geomean_conf for latex tables:
Inspired by this post: # https://stackoverflow.com/questions/44325464/how-to-control-knitr-kable-scientific-notation
}
