% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean function}
\usage{
gm_mean(x, na.rm = TRUE, offset = 0, zero.propagate = FALSE, conf.level = NA)
}
\arguments{
\item{x}{numeric vector as input}

\item{na.rm}{logical, whether missing values are allowed.}

\item{offset}{Optional variable: adds a value to the data (1 is often used), so that no more zeros occur.}

\item{zero.propagate}{if TRUE, the output will be zero if any value is zero. Overrides offset.}

\item{conf.level}{If not missing, the confidence interval will be given together with the geometric mean}
}
\value{
a vector of length 1 or, if conf.level is not missing, length 3
}
\description{
Calculates the geometric mean.
If conf.level is given, a vector will be output containing the geometric mean,
the lower bound confidence value, and the upper bound confidence value.
Loosely based on https://stackoverflow.com/a/25555105/11856430
}
\examples{
gm_mean(c(1,4,2,3))
}
