% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_data}
\alias{immune_data}
\title{Toy data of immune cell subset concentrations}
\format{
An object of class \code{data.frame} with 100 rows and 23 columns.
}
\source{
https://doi.org/10.3389/fimmu.2021.670070
}
\usage{
immune_data
}
\description{
Toy data of immune cell subset concentrations
}
\keyword{datasets}
