% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{mean.plus.range}
\alias{mean.plus.range}
\title{Function for table output giving mean plus range:}
\usage{
\method{mean}{plus.range}(x, digits = 1, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Function for table output giving mean plus range:
}
