% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{mean.sd}
\alias{mean.sd}
\title{older function used for nice table output}
\usage{
\method{mean}{sd}(var, n.digits = 2, na.rm = T)
}
\arguments{
\item{na.rm}{}
}
\value{

}
\description{
older function used for nice table output
}
