% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mean_plus_range}
\alias{mean_plus_range}
\title{Function for table output giving mean plus range:}
\usage{
mean_plus_range(x, digits = 1, ...)
}
\arguments{
\item{x}{numerical value}

\item{digits}{number of digits to use}

\item{...}{other vars, parsed to mean/min/max/sd functions}
}
\value{
character string of mean plus range
}
\description{
Function for table output giving mean plus range:
}
