% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mean_sd}
\alias{mean_sd}
\title{Mean plus standard deviation}
\usage{
mean_sd(var, na.rm = T)
}
\arguments{
\item{var}{variable name}

\item{na.rm}{logical, indicating whether NA values should be excluded}
}
\value{
Character string
}
\description{
A vector will be output containing the arithmetic mean and the standard
deviation. The output can be used directly in the function
\code{\link{format_values}} to create summary tables.
}
\examples{
mean_sd(rnorm(100, 1000, 10))
}
