% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mean_sd_range}
\alias{mean_sd_range}
\title{Mean, SD and range}
\usage{
mean_sd_range(x, digits = 1, ...)
}
\arguments{
\item{x}{numerical value}

\item{digits}{number of digits to use}

\item{...}{other vars, parsed to mean/min/max/sd functions}
}
\value{
character string of mean plus range
}
\description{
Function for table output giving mean, standard deviation,
plus range (minimum value and maximum value). Can be useful
in summary tables.
}
\examples{
mean_sd_range(rnorm(1000, 1000, 100))
}
