% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{median_iqr}
\alias{median_iqr}
\title{Table output helper}
\usage{
median_iqr(var, n.digits = 2, na.rm = T)
}
\arguments{
\item{var}{variable name}

\item{n.digits}{number of digits to use in output}

\item{na.rm}{exclude NA's or not}
}
\value{
Character string
}
\description{
older function used for nice table output
}
