% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{perc_number}
\alias{perc_number}
\title{Percentage and number in ordinal variable.}
\usage{
perc_number(var, count.condition, n.digits = 1, na.rm = TRUE)
}
\arguments{
\item{var}{character of factor vector}

\item{count.condition}{character string. Variable to group by}

\item{n.digits}{number of digits to use in output}

\item{na.rm}{logical, indicating whether NA values should be excluded}
}
\value{
Character string
}
\description{
Helper function to create summary statistics of ordinal or categorical
variables, for use in a summary table.
It will display the occurrence of a condition within an ordinal or
categorical variable. Works best with two groups.
}
\examples{
groupvar <-  sample(c("Men", "Women"), 80, TRUE)
perc_number(groupvar, count.condition = "Men")
}
