% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{perform_single_pair_test}
\alias{perform_single_pair_test}
\title{perform paired testing of associations.}
\usage{
perform_single_pair_test(
  single.pair,
  data.set,
  resampling.number = 10^4,
  stratum = NULL
)
}
\arguments{
\item{single.pair}{character varlue of length 2}

\item{data.set}{data frame to use}

\item{resampling.number}{number resamplings (see also coin::approximate)}

\item{stratum}{character value. blocking vars to use}
}
\description{
Note: deprecated function. This function can be completely replaced by test_association.
Only thing missing is the check of blocks, whether blocks with n <= 2
observations exists.
}
