% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFanalysis.R
\name{RFanalysis}
\alias{RFanalysis}
\title{Function to perform Random Forest analysis}
\usage{
RFanalysis(
  data,
  subset = NULL,
  predictors,
  confounders,
  dependent.var,
  ntree = 10000,
  select.best.mtry = TRUE,
  ...
)
}
\arguments{
\item{data}{dataframe to use (original: ISA.cellular)}

\item{subset}{whether or not to use subset of data}

\item{predictors}{character value, predictors (original: cell.subsets.to.analyse)}

\item{confounders}{character value, confounding variables (original: confounding.vars)}

\item{dependent.var}{character value (original: dependent.variable)}

\item{ntree}{numerical, number of trees to use in random forest}

\item{select.best.mtry}{logical}

\item{...}{other parameters parsed to the randomForest function}
}
\value{
a randomForest object
}
\description{
Note: dataset is expected to be in "wide" format for this function.
Note2: This function is a work in progress and still needs some more testing.
Use with caution.
}
