% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{scientific_10}
\alias{scientific_10}
\title{Create scientific notation in ggplot2 functions}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{any numeric value that should have scientific notation}
}
\value{
value in scientific notation
}
\description{
Create scientific notation in ggplot2 functions
}
