% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_studies.R
\name{test_association}
\alias{test_association}
\title{Test association}
\usage{
test_association(
  dataset,
  response.var,
  explanatory.var,
  stratum = NULL,
  n.resample = 10^4,
  ...
)
}
\arguments{
\item{dataset}{dataset (data frame format, original: data.wide)}

\item{response.var}{response variable (for example Granulocytes_count)}

\item{explanatory.var}{explanatory variable (for example Frailty.index)}

\item{stratum}{blocking variables (character string)}

\item{n.resample}{number of resamples}

\item{...}{other variables that influence the coin::approximate function}
}
\value{
data frame with results
}
\description{
Function that can be used to test an association between two variables.
It uses permutation testing functions from the coin package.
It gives a consistent data frame output.
}
\details{
Note: Errors can occur when some groups within the blocking variables
contain less than two observations.
}
