% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variable_pair_combinations}
\alias{variable_pair_combinations}
\title{Create all combinations of variable names}
\usage{
variable_pair_combinations(data, sel = tidyselect::everything())
}
\arguments{
\item{data}{a data frame as input}

\item{sel}{parsed to dplyr::select(). To include/exclude variables.}
}
\value{
a character value matrix with two columns, containing all possible pairs of variable names
}
\description{
Useful in a an apply loop when associations should be tested
between many variables in a data frame.
}
\examples{
variable_pair_combinations(immune_data, sel = -c(Sex, Frailty.index, Batch))
}
