% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{weighted.average.rho}
\alias{weighted.average.rho}
\title{Title}
\usage{
weighted.average.rho(data = data.for.test.i, x, y, stratum = "block")
}
\arguments{
\item{data}{expects a data frame}

\item{x}{Character value, first variable. Should be in data}

\item{y}{Character value, second variable. Should be in data}

\item{stratum}{stratum (factor) to block the association with}
}
\value{
numeric value (weighted average rho)
}
\description{
Title
}
\examples{
weighted.average.rho(iris, "Petal.Length", "Petal.Width", "Species")

}
