% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_average_rho.R
\name{weighted_average_rho}
\alias{weighted_average_rho}
\title{Weighted average rho}
\usage{
weighted_average_rho(data, x, y, stratum = "block")
}
\arguments{
\item{data}{expects a data frame}

\item{x}{Character value, first variable. Should be in data}

\item{y}{Character value, second variable. Should be in data}

\item{stratum}{stratum (factor) to block the association with}
}
\value{
numeric value (weighted average rho)
}
\description{
calculates weighted average rho in a blocked correlation study, as discussed
with Jose Ferreira (RIVM)
}
\examples{
weighted_average_rho(iris, "Petal.Length", "Petal.Width", "Species")

}
