% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_studies.R
\name{association_study}
\alias{association_study}
\title{Perform association study}
\usage{
association_study(data, expl_var_names, response.var, ...)
}
\arguments{
\item{data}{data frame to use}

\item{expl_var_names}{character value of all column names that need to be compared to the response
variable. This will we looped over the variable "explanatory.var" in the
function \code{\link{test_association}}}

\item{response.var}{response variable. Will be parsed to  \code{\link{test_association}}}

\item{...}{other values will be parsed to \code{\link{test_association}}}
}
\value{
data frame with results as output
}
\description{
This function is a small wrapper around the function \code{\link{test_association}}.
It tests associations between an explanatory variable and multiple response
variables. The response variables are columns in a data frame and their
colum names should be given in a character string.
}
\details{
Note: Response variables should be of the same type
(e.g. all numerical, all categorical).
Note 2: all blocks should contain enough observations.
}
