% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_studies.R
\name{association_study}
\alias{association_study}
\title{Association studies}
\usage{
association_study(
  subset.parameter,
  subset.column,
  dataset,
  numerical.value,
  comparison.value,
  stratum = NULL,
  n.resample = 10^4,
  ...
)
}
\arguments{
\item{subset.parameter}{character value that can be found in the subset.column (original: Granulocytes_Count)}

\item{subset.column}{column containing subset names (original: Cellname)}

\item{dataset}{data frame to use (original: data.long)}

\item{numerical.value}{a numerical value for the comparison}

\item{comparison.value}{the second parameter. If it is a factor, the wilcoxon test will be used. If it is numerical, the spearman_test will be used}

\item{stratum}{character string. Can be one or several blocking variables.}

\item{n.resample}{number of random samples for permutation testing}

\item{...}{Other parameters that influence the function coin::approximate}
}
\value{
data frame with results
}
\description{
Function to perform associations between two parameters
Associations with p values calculated by permutation, using coin package
Note: dataset is expected to be in "long" format for this function.
Note2: missing values are (silently) deleted
Use the function in an apply loop (or map_dfr loop) to perform it on
many different subsets in the data.
}
\examples{
cell_vars <- cell_vars <- names(immune_data)[!names(immune_data) \%in\%
c("Sex", "Frailty.index", "Batch")]
immune_data_long <- immune_data \%>\%
tidyr::pivot_longer(names_to = "cellnames", values_to = "value", cell_vars)

purrr::map_dfr(cell_vars, .f = ~association_study(
dataset = immune_data_long,
subset.parameter = .,
subset.column = "cellnames",
numerical.value = "value",
comparison.value = "Frailty.index",
stratum = c("Sex", "Batch")
))

}
