% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_studies.R
\name{association_study}
\alias{association_study}
\title{Perform association study}
\usage{
association_study(
  data,
  response.var = NULL,
  vars.to.select = tidyselect::everything(),
  stratum = NULL,
  ...
)
}
\arguments{
\item{data}{data frame to use}

\item{response.var}{Character value, optional. Response variable. If provided, associations will
only be tested for this single response variable.}

\item{vars.to.select}{Variables to select. Will be used within dplyr::select.
All possible pairs of associations between these variables will be tested.
Standard all variables will be selected.}

\item{stratum}{character vector. When used, the tests will be blocked by these variables.}

\item{...}{other parameters that will be parsed to \code{\link{test_association}}.}
}
\value{
data frame with results as output
}
\description{
This function is a wrapper around the function \code{\link{test_association}}.
This function tests associations between multiple variables by repeatedly
calling the function \code{\link{test_association}}.
Standard, all possible single associations between pairs of variables in the
given data frame are tested.
}
\details{
The response variables are columns in a data frame and their
column names should be given in a character string.

Note 2: Associations can be stratified (see options for
\code{\link{test_association}}). All strata/blocks should contain enough
observations.
}
\examples{
## Test pairwise associations between all variables in a a data frame:
association_study(immune_data[, 1:10])

## Test associations of a response variable with all other variables in a data frame:
association_study(immune_data, "Frailty.index")

## Test associations with a selection of variables:
association_study(immune_data, "Frailty.index", c(Tregs, Neutrophils))

## Excluding variables is also possible:
association_study(immune_data, "Frailty.index", -c(Tregs, Neutrophils))

## as well as using a blocked design:
association_study(immune_data, "Frailty.index", -c(Tregs, Neutrophils),
stratum = c("Batch", "Sex"))

}
\seealso{
\code{\link{test_association}} for further arguments that can be used (for
example, changing the resampling number n.resample),
\code{\link{BH_selection}} for applying the Benjamini-Hochberg selection
procedure based on False Discovery Rate.
}
