% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_study_functions.R
\name{association_study_long}
\alias{association_study_long}
\title{Association studies in long format}
\usage{
association_study_long(data, response.names, ...)
}
\arguments{
\item{data}{data frame to use}

\item{response.names}{character value of column that contains all the names of the response values that need to be investigated separately}

\item{...}{other values that will be parsed to the function association_study_wide}
}
\value{
data frame with results as output
}
\description{
This function, in combination with association_study_wide,
can basically replace the function 'association_study'.
While I used the function association_study in my publications,
I think this implementation is better. The function association_study_wide
is more intuitive, and this function is just a small wrapper
that you can use when data is in long format.
}
