% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_studies.R
\name{association_study_long}
\alias{association_study_long}
\title{Perform association study}
\usage{
association_study_long(data, expl_var_names, ...)
}
\arguments{
\item{data}{data frame to use}

\item{expl_var_names}{character value of column that contains all the names of the response values that need to be investigated separately}

\item{...}{other values will be parsed to \code{\link{test_association}}}
}
\value{
data frame with results as output
}
\description{
This function, in combination with \code{\link{test_association}},
can basically replace the function '\code{\link{association_study}}'.
I think \code{\link{test_association}} is more intuitive,
and this function is just a small wrapper
that you can use when data is in long format, mimicking behavior
of \code{\link{association_study}}.
}
\details{
Note: Response variables should be of the same type
(e.g. all numerical, all categorical).
Note 2: all blocks should contain enough observations.
}
