% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_study_functions.R
\name{associations_per_ID}
\alias{associations_per_ID}
\title{This function will investigate the association between biomarkers over time
(longitudinal analyses) with ID/individual as blocking variable}
\usage{
associations_per_ID(
  x,
  dataset = immune.trajectories.long,
  comparison.var = "flowratio",
  nresample = 10^5
)
}
\arguments{
\item{nresample}{}
}
\value{

}
\description{
This function will investigate the association between biomarkers over time
(longitudinal analyses) with ID/individual as blocking variable
}
