% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{format_summary_values}
\alias{format_summary_values}
\title{Format values for table input}
\usage{
format_summary_values(
  x,
  n.digits = 3,
  na.rm = FALSE,
  conf.level = 0.95,
  latex_output = TRUE
)
}
\arguments{
\item{x}{numeric values that need to be summarized (for now, only geometric mean + confidence interval is supported)}

\item{n.digits}{digits to preserve in output}

\item{na.rm}{whether or not to remove NA's}

\item{conf.level}{confidence interval, standard 0.95}

\item{latex_output}{Logical, Whether dollar signs should be used around the output string,
for latex formatting in knitr::kable. Make sure to set escape=FALSE in
knitr::kable function when using this.}
}
\value{
Character value to input in kable function
}
\description{
Formats the results of gm_mean so that it can be input into
a (latex) table, such as the output of knitr::kable().
Aims to give a publication-ready result
}
\details{
Inspired by this post: # https://stackoverflow.com/questions/44325464/how-to-control-knitr-kable-scientific-notation
}
\examples{
data <- abs(rnorm(n = 100, mean = 10000, sd = 10000))
format_summary_values(data)

}
