% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_data}
\alias{immune_data}
\title{Example data of immune cell subset concentrations.
The variables "Frailty.index", "Sex", and "Batch" are random numbers.
The other numbers and names are a selection of data from the source mentioned
below.#'}
\format{
An object of class \code{data.frame} with 100 rows and 23 columns.
}
\source{
https://doi.org/10.3389/fimmu.2021.670070
}
\usage{
immune_data
}
\description{
Example data of immune cell subset concentrations.
The variables "Frailty.index", "Sex", and "Batch" are random numbers.
The other numbers and names are a selection of data from the source mentioned
below.#'
}
\keyword{datasets}
