% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{mean_sd}
\alias{mean_sd}
\title{Table output helper}
\usage{
mean_sd(var, n.digits = 2, na.rm = T)
}
\arguments{
\item{var}{variable name}

\item{n.digits}{number of digits to round number}

\item{na.rm}{exclude NA's or not}
}
\value{
Character string
}
\description{
older function used for nice table output
}
