% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{median_iqr}
\alias{median_iqr}
\title{Median + interquartile range}
\usage{
median_iqr(var, na.rm = TRUE)
}
\arguments{
\item{var}{numerical vector}

\item{na.rm}{logical, indicating whether NA values should be excluded}
}
\value{
Character string
}
\description{
Creates a vector with the mean and interquartile range.
Output can be used directly in the function
\code{\link{format_values}} to create summary tables.
}
\examples{
median_iqr(rnorm(100, 1000, 100))
}
