% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{pagebreak}
\alias{pagebreak}
\title{Function to add pagebreaks in Rmarkdown files.
For more info see: https://stackoverflow.com/a/55064070}
\usage{
pagebreak()
}
\value{
either a html or a latex tag that adds a page break
}
\description{
Function to add pagebreaks in Rmarkdown files.
For more info see: https://stackoverflow.com/a/55064070
}
