% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{perc_number}
\alias{perc_number}
\title{Table output helper}
\usage{
perc_number(var, count.condition, n.digits = 1, na.rm = T)
}
\arguments{
\item{var}{variable name}

\item{count.condition}{factor variable to group by}

\item{n.digits}{number of digits to use in output}

\item{na.rm}{exclude NA's or not}
}
\value{
Character string
}
\description{
older function used for nice table output
}
