% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_study_functions.R
\name{plot_p_histogram}
\alias{plot_p_histogram}
\title{Function to plot P histogram}
\usage{
plot_p_histogram(
  data,
  p_value = "p.value",
  title.to.plot = "P-values of the association study",
  gamma.hat = 1
)
}
\arguments{
\item{data}{dataframe to use}

\item{p_value}{character string. name of column containing p values}

\item{title.to.plot}{title of the plot to produce}

\item{gamma.hat}{gamma hat to use}
}
\value{
histogram
}
\description{
Function to plot P histogram
}
