% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFanalysis.R
\name{RFanalysis}
\alias{RFanalysis}
\title{Function to perform Random Forest analysis
Note: dataset is expected to be in "wide" format for this function.}
\usage{
RFanalysis(
  data = ISA.cellular,
  subset = NULL,
  predictors = cell.subsets.to.analyse,
  confounders = confounding.vars,
  dependent.var = dependent.variable,
  ntree = 10000,
  select.best.mtry = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Function to perform Random Forest analysis
Note: dataset is expected to be in "wide" format for this function.
}
