% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scientific_10}
\alias{scientific_10}
\title{Scientific notation}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{any numeric value that should have scientific notation}
}
\value{
value in scientific notation
}
\description{
Create scientific notation to use in ggplot2 functions. Inspired by
\url{https://stackoverflow.com/a/18530540/11856430}
}
